+++ nail/CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
+- added jemalloc support with feature flag `jemalloc`
 - added `Stats` module for keeping track of pipeline summary statistics
 - search pipeline now uses the `thread_local` crate instead of letting Rayon clone whenever it wants to
 - added progress bars